[REFACTOR] Likes and OAuth controllers
