+++ server/controllers/Like/Index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  post: require('./Post.js'),
+  delete: require('./Delete.js'),
+};
