+++ server/controllers/OAuth/Index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  githubLogin: require('./GitHubLogin.js'),
+  kakaoLogin: require('./KakaoLogin.js'),
+};
