+++ server/routes/OAuth.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const oauthController = require('../controllers/OauthController');
+const oauthController = require('../controllers/OAuth/Index.js');
 
 router.post('/github', oauthController.githubLogin);
 router.post('/kakao', oauthController.kakaoLogin);
